Coding from creating tables to writing DDL, DML commands in MS SQL Server
In the previous semester, I and my group did a project called "Esenler Airport"  The task was to think of an imaginary company and create at least 10 main tables in its database in Microsoft SQL Server. The project had 2 phases and the want list of these phases was given to us by our Teacher. The first phase contained schemas and graphics that shows us the structure of the tables which was basically the E-R Table. In this ER Table, we used the attributes that could be shown in a real database in a real airport company. We decided which data types we should be using for each attribute and explained why we used them in detail in our documentation. In our second phase, we actually coded the tables and attributes in MS SQL Server.  We created meaningful select statements for reporting which included different aggregate functions, joins, group by, having, exist, in, union, intersect, except clauses, created meaningful indexes, views, stored procedures, transactions and used all DDL statements. The task could be made in German or English, we chose English to use in our documentation.
We gave this project most of our time and importance more than exams because we thought this project was our first opportunity to gain real experience and knowledge about coding other than the theoretical lectures.
To me in order to be successful in a project comes with investing time and effort rapidly and wisely as well as being open to discussing the problems with the team frequently.
We did this and got full marks for our project.